requirements.txt
streamlit==1.38.0
feedparser==6.0.11
python-dotenv==1.0.1
pyyaml==6.0.2
langchain==0.2.16
langchain-openai==0.1.23
scikit-learn==1.5.2
numpy==2.1.1
